Completed sections 9 and 10
Added rental view for customers to rent movies and track available movies
from stock. This required extending the move model to have a count of available
inventory. The problem needed to be structured first, then the back end
was built, then the front end.

Added dependencies to Glimpse, toastr, and typeahead